+++ Projecto1/src/io/altar/jseproject/textinterface/states/MenuCreateProduct.java
@@ -7,11 +7,11 @@ public class MenuCreateProduct extends State {
 	@Override
 	public int show() {
 		System.out.println("Inserir desconto");
-		double desconto = scu.getDouble();
+		int desconto = scu.getInt();
 		System.out.println("Inserir iva");
 		int iva = scu.getIva();
 		System.out.println("Inserir preço");
-		int preco = scu.getInt();
+		double preco = scu.getDouble();
 		Product p = new Product(desconto,iva,preco);
 		prodRep.createEntity(p);
 		System.out.println("--- Produto criado ---");
