+++ Projecto1/src/io/altar/jseproject/textinterface/states/MenuEditShelf.java
@@ -12,11 +12,11 @@ public class MenuEditShelf extends State {
 		Shelf sh = shelfRep.getEntityById(id);
 		Product p = sh.getPr();
 		System.out.println("Inserir desconto do produto");
-		p.setDiscount(scu.getDouble());
+		p.setDiscount(scu.getInt());
 		System.out.println("Inserir iva do produto");
 		p.setIva(scu.getInt());
 		System.out.println("Inserir preço do produto");
-		p.setPrice(scu.getInt());
+		p.setPrice(scu.getDouble());
 		System.out.println("Inserir capacidade");
 		sh.setCapacity(scu.getInt());
 		System.out.println("Inserir preço diário de aluguer");
